Auth and Products Controller Created Auth Email Verification Completed
