+++ BACKEND/Models/Token.Model.js
@@ -2,11 +2,10 @@ const mongoose = require("mongoose");
 
 const tokenSchema = mongoose.Schema(
   {
-    ID: {
+    authID: {
       type: mongoose.Schema.Types.ObjectId,
       required: true,
       ref: "auth",
-      unique: true,
     },
     token: { type: String, required: true },
     createdAt: { type: Date, default: Date.now(), expires: 3600 },
@@ -18,6 +17,4 @@ const tokenSchema = mongoose.Schema(
 
 const TokenModel = mongoose.model("token", tokenSchema);
 
-module.exports = {
-  TokenModel,
-};
+module.exports = TokenModel;
