+++ BACKEND/index.js
@@ -6,20 +6,20 @@ const connection = require("./Config/db");
 const ProductsRouter = require("./Routes/Products.Router");
 const AuthRouter = require("./Routes/Auth.Router");
 const ErrorMiddleware = require("./Middleware/Error.Middleware");
-// const fileUpload = require("express-fileupload");
 const AccessControlMiddleware = require("./Middleware/AccessControl.Middleware");
 
 const app = express();
 app.use(AccessControlMiddleware);
 app.use(express.json());
 app.use(express.urlencoded({ extended: false }));
-app.use("/uploads", express.static("uploads"))
+app.use("/uploads", express.static("uploads"));
 app.use("/products", ProductsRouter);
 app.use("/auth", AuthRouter);
 app.use(ErrorMiddleware);
+app.set("view engine", "hbs");
 
 app.get("/", (req, res) => {
-  res.status(200).send("Welcome To Subha Mukherjee's E-Commerce");
+  res.status(200).send("Welcome To ShopKart");
 });
 
 app.listen(port, async () => {
