+++ BACKEND/views/error.hbs
@@ -1,0 +1,55 @@
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>404 Error</title>
+  </head>
+  <style>
+    body {
+        display: flex;
+        justify-content: center;
+        align-items: center;
+        background-color: black;
+    }
+
+    div {
+        width: 40%;
+        text-align: center;
+    }
+
+    img {
+        width: 100%;
+        height: auto;
+    }
+
+    .green_btn {
+        margin-top: 10px;
+        border: none;
+        outline: none;
+        padding: 12px 0;
+        background-color: #3bb19b;
+        border-radius: 20px;
+        width: 180px;
+        font-weight: bold;
+        font-size: 14px;
+        cursor: pointer;
+    }
+
+    a {
+      text-decoration: none;
+    }
+  </style>
+  <body>
+    <div>
+      <img
+        src="https://media.tenor.com/IHdlTRsmcS4AAAAC/404.gif"
+        alt="404 Not Found"
+      />
+
+      <Link to="/login">
+          <button class="green_btn"><a href="https://prdb-project.vercel.app">Home</a></button>
+      </Link>
+    </div>
+  </body>
+</html>
