+++ BACKEND/views/index.hbs
@@ -1,0 +1,50 @@
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Success</title>
+  </head>
+  <style>
+    body {
+        margin: 0;
+        padding: 0;
+        overflow-x: hidden;
+        box-sizing: border-box;
+    }
+
+    .container {
+        width: 100vw;
+        height: 100vh;
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        flex-direction: column;
+    }
+
+    .green_btn {
+        margin-top: 10px;
+        border: none;
+        outline: none;
+        padding: 12px 0;
+        background-color: #3bb19b;
+        border-radius: 20px;
+        width: 180px;
+        font-weight: bold;
+        font-size: 14px;
+        cursor: pointer;
+    }
+  </style>
+
+  <body>
+    <div class="container">
+          <video loop autoplay muted>
+            <source src="https://i.gifer.com/7efs.mp4" />
+          </video>
+          <h1>Email verified successfully</h1>
+          <Link to="/login">
+            <button class="green_btn"><a href="https://prdb-project.vercel.app/login">Login</a></button>
+          </Link>
+        </div>
+  </body>
+</html>
